% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adapt_list}
\alias{adapt_list}
\title{Adapt the path of class \code{url}}
\usage{
adapt_list(old_path, path)
}
\arguments{
\item{old_path}{Passed path in parameter \code{domain}}

\item{path}{Path of API endpoint}
}
\value{
Returns a vector of the resulting path.
}
\description{
Adapt the path of class \code{url} to consider the old path when modifying
}
\section{Warning}{

Internal function
}

