% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{basic_issues_info}
\alias{basic_issues_info}
\title{Extract the basic key information of the issues}
\usage{
basic_issues_info(x)
}
\arguments{
\item{x}{JIRA issue list item.}
}
\value{
Returns \code{data.frame} with basic field information.
}
\description{
Internal function to extract the basic key information
as a \code{data.frame}.
}
\section{Warning}{

Internal function
}

