% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{basic_jql_fields}
\alias{basic_jql_fields}
\title{Returns default JQL fields used}
\usage{
basic_jql_fields()
}
\value{
Returns a \code{character} vector with the default JQL fields.
}
\description{
Internal function used to define the default JQL fields used for the \code{get_jira_issues()} function.
}
\section{Warning}{

Internal function
}

\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
