% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{choose_field_function}
\alias{choose_field_function}
\title{Function to choose for the right field parser function}
\usage{
choose_field_function(x, type)
}
\arguments{
\item{x}{The fields nested data to flatten.}

\item{type}{The fields' name.}
}
\value{
Returns a parsed, cleaned \code{data.frame} with all the fields.
}
\description{
Internal function to choose/switch to the correct function to parse each field for each issue
}
\section{Warning}{

Internal function
}

\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
