% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{conc}
\alias{conc}
\title{Concatenates multiple strings}
\usage{
conc(x, y = ",", decr = FALSE, unique = TRUE)
}
\arguments{
\item{x}{A single character vector to concatenate together.}

\item{y}{By default a \code{,} string used to define the character to collapse the \code{x} parameter.}

\item{decr}{Optional logical parameter that defines the sorting order, by default set to \code{FALSE}, which results in an alphabetical order.}

\item{unique}{Optional logical parameter to concatenate only unique values, by default set to \code{TRUE}}
}
\value{
Returns a single character string.
}
\description{
Internal function with an opinionated default behaviour to concatenate charcter values.
}
\section{Warning}{

Internal function
}

\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
