% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{error_response}
\alias{error_response}
\title{Function to inform the user about possible error codes}
\usage{
error_response(x)
}
\arguments{
\item{x}{REST API response status error code.}
}
\value{
Returns a single character string with an error message.
}
\description{
Internal function show potential error code messages to the user.
}
\section{Warning}{

Internal function
}

\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
