% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{get_jira_credentials}
\alias{get_jira_credentials}
\title{Retrieves the previously saved JIRA credentials}
\usage{
get_jira_credentials()
}
\value{
Returns a \code{data.frame} with the saved JIRA credentials
}
\description{
Retrieves a \code{data.frame} with the JIRA credentials previously saved into the environment under the JIRAGILER_PAT variable through the \code{save_jira_credentials()} function.
}
\examples{
\dontrun{
save_jira_credentials(domain="https://bugreports.qt.io",
                      username='__INSERT_YOUR_USERNAME_HERE__',
                      password='__INSERT_YOUR_PASSWORD_HERE__')
get_jira_credentials()
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
