% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{get_jira_groups}
\alias{get_jira_groups}
\title{Get JIRA server groups \code{data.frame}}
\usage{
get_jira_groups(
  domain = NULL,
  username = NULL,
  password = NULL,
  verbose = FALSE,
  maxResults = 1000
)
}
\arguments{
\item{domain}{Custom JIRA domain URL as for example \href{https://bitvoodoo.atlassian.net}{https://bitvoodoo.atlassian.net}. Can be passed as a parameter or can be previously defined through the \code{save_jira_credentials()} function.}

\item{username}{Username used to authenticate the access to the JIRA \code{domain}. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.}

\item{password}{Password used to authenticate the access to the JIRA \code{domain}. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.}

\item{verbose}{Explicitly informs the user of the JIRA API request process.}

\item{maxResults}{Number of maximum groups to return. Set by default to \code{1000}.}
}
\value{
Returns a \code{data.frame} with all the JIRA groups
}
\description{
Makes a request to JIRA's latest REST API to retrieve all the groups in JIRA.
}
\section{Warning}{

The function works with the JIRA REST API. Thus, to work it needs an internet connection. Calling the function too many times might block your access and you will have to access manually online and enter a CAPTCHA at \href{https://jira.yourdomain.com/secure/Dashboard.jspa}{jira.yourdomain.com/secure/Dashboard.jspa}.
}

\examples{
get_jira_groups("https://bitvoodoo.atlassian.net")
}
\seealso{
For more information about Atlassians JIRA API go to \href{https://docs.atlassian.com/software/jira/docs/api/REST/8.3.3/}{JIRA API Documentation}
}
\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
