% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{get_jira_projects}
\alias{get_jira_projects}
\title{Retrieves all projects as a \code{data.frame}}
\usage{
get_jira_projects(
  domain = NULL,
  username = NULL,
  password = NULL,
  expand = NULL,
  verbose = FALSE
)
}
\arguments{
\item{domain}{Custom JIRA domain URL as for example \href{https://bitvoodoo.atlassian.net}{https://bitvoodoo.atlassian.net}. Can be passed as a parameter or can be previously defined through the \code{save_jira_credentials()} function.}

\item{username}{Username used to authenticate the access to the JIRA \code{domain}. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.}

\item{password}{Password used to authenticate the access to the JIRA \code{domain}. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.}

\item{expand}{Specific JIRA fields the user wants to obtain for a specific field. Optional parameter.}

\item{verbose}{Explicitly informs the user of the JIRA API request process.}
}
\value{
Returns a \code{data.frame} with a list of projects for which the user has the BROWSE, ADMINISTER or PROJECT_ADMIN project permission.
}
\description{
Makes a request to JIRA's latest REST API to retrieve all projects and their basic project information (Name, Key, Id, Description, etc.).
}
\section{Warning}{

The function works with the latest JIRA REST API and to work you need to have a internet connection. Calling the function too many times might block your access, you will receive a 403 error code. To unblock your access you will have to access interactively through your browser, signing out and signing in again, and might even have to enter a CAPTCHA at \href{https://jira.yourdomain.com/secure/Dashboard.jspa}{jira.enterprise.com/secure/Dashboard.jspa}. This only happens if the API is called upon multiple times in a short period of time.
}

\examples{
\dontrun{
get_jira_projects("https://bitvoodoo.atlassian.net")
}
}
\seealso{
For more information about Atlassians JIRA API go to \href{https://docs.atlassian.com/software/jira/docs/api/REST/8.3.3/}{JIRA API Documentation}
}
\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
