% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JirAgileR-package.R
\docType{package}
\name{JirAgileR-package}
\alias{JirAgileR}
\alias{JirAgileR-package}
\title{JirAgileR: JIRA REST API Wrapper for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Allows to interact with the 'JIRA SERVER REST API' to analyze the retrieved data in R. For further information about the API visit <https://docs.atlassian.com/software/jira/docs/api/REST/8.9.1/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matbmeijer.github.io/JirAgileR/}
  \item \url{https://github.com/matbmeijer/JirAgileR}
  \item Report bugs at \url{https://github.com/matbmeijer/JirAgileR/issues}
}

}
\author{
\strong{Maintainer}: Matthias Brenninkmeijer \email{matthias@brennotten.net}

}
\keyword{internal}
