% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{Project2R}
\alias{Project2R}
\title{Obtain project issues as \code{df}}
\usage{
Project2R(dn, user = NULL, password = NULL, project, search = "name")
}
\arguments{
\item{dn}{Custom JIRA domain URL as for example \href{https://bitvoodoo.atlassian.net}{https://bitvoodoo.atlassian.net}}

\item{user}{Username used to authenticate access to JIRA your domain. If no username and password is given no authentication is made.}

\item{password}{Password used to authenticate access to JIRA domain. If no username and password is given no authentication is made.}

\item{project}{Exact name of the JIRA project e.g. 'Package Development'.}

\item{search}{Currently defaults to \code{"name"}. In future the objective is to be able to search by name and ID of the project or even single issue.}
}
\value{
Returns a formatted \code{data.frame} with all issues as rows and with the default columns 'Key', 'Status', 'Version', 'Assignee_Name', 'Assignee_Email', 'Description', 'Created', 'Updated', 'Summary', 'Priority' and 'Type'.
}
\description{
Calls JIRA's v2 Rest API with basic authentication to get all issues of a JIRA project.
}
\section{Warning}{

The function works with the JIRA REST v2 API and to workyou need to have a internet connection. Calling the function too many times might block your access and you will have to access manually online and enter a CAPTCHA at \href{https://jira.yourdomain.com/secure/Dashboard.jspa}{jira.enterprise.com/secure/Dashboard.jspa}
}

\examples{
Project2R("https://bitvoodoo.atlassian.net", project="Congrats for Confluence", search="name")
}
\seealso{
For more information about Atlassians JIRA API go to \href{https://docs.atlassian.com/software/jira/docs/api/REST/7.6.1/}{https://docs.atlassian.com/software/jira/docs/api/REST/7.6.1/}
}
\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{Github}
}
