% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project2R.R
\name{supported_jql_fields}
\alias{supported_jql_fields}
\title{Returns the supported JQL fields}
\usage{
supported_jql_fields()
}
\value{
Returns a character vector of all the supported JQL fields.
}
\description{
Function shows all the supported JQL fields that are available to choose for the \code{JiraQuery2R()} function.
}
\author{
Matthias Brenninkmeijer \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
