% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_date}
\alias{to_date}
\title{Transforms JIRA date character to POSIXlt format}
\usage{
to_date(x)
}
\arguments{
\item{x}{Character vector to transform into a \code{POSIXlt}.}
}
\value{
Returns a \code{POSIXlt} object vector.
}
\description{
Internal function to parse the date from JIRA character vectors.
}
\section{Warning}{

Internal function
}

