% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JirAgileR-package.R
\docType{package}
\name{JirAgileR-package}
\alias{JirAgileR}
\alias{JirAgileR-package}
\title{JirAgileR: JIRA REST API Wrapper for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Allows to interact with the JIRA REST API to analyze the data in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matbmeijer.github.io/JirAgileR}
  \item \url{https://github.com/matbmeijer/JirAgileR}
}

}
\author{
\strong{Maintainer}: Matthias Brenninkmeijer \email{matthias@brennotten.net}

}
\keyword{internal}
