% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JirAgileR-package.R
\docType{package}
\name{JirAgileR-package}
\alias{JirAgileR}
\alias{JirAgileR-package}
\title{JirAgileR: JIRA REST API Wrapper for R}
\description{
Allows you to interact with JIRA REST API to analyze the data in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matbmeijer/JirAgileR}
}

}
\author{
\strong{Maintainer}: Matthias Brenninkmeijer \email{matthias@brennotten.net}

}
\keyword{internal}
